Move shared code into one package
