+++ apps/parsley/package.json
@@ -40,8 +40,7 @@
     "@apollo/client": "3.7.12",
     "@emotion/react": "11.11.0",
     "@emotion/styled": "11.11.5",
-    "@evg-ui/components": "*",
-    "@evg-ui/utils": "*",
+    "@evg-ui/lib": "*",
     "@leafygreen-ui/badge": "8.0.2",
     "@leafygreen-ui/button": "21.0.6",
     "@leafygreen-ui/card": "10.0.5",
@@ -89,6 +88,7 @@
   "devDependencies": {
     "@emotion/babel-plugin": "11.11.0",
     "@evg-ui/eslint-config": "*",
+    "@evg-ui/lint-staged": "*",
     "@graphql-codegen/cli": "5.0.2",
     "@graphql-codegen/typescript": "4.0.1",
     "@graphql-codegen/typescript-operations": "4.0.1",
