+++ apps/parsley/src/hooks/useFilterParam/index.ts
@@ -1,5 +1,5 @@
 import { useCallback } from "react";
-import { conditionalToArray } from "@evg-ui/utils";
+import { conditionalToArray } from "@evg-ui/lib/utils/array";
 import { QueryParams } from "constants/queryParams";
 import { useQueryParams } from "hooks/useQueryParam";
 import { Filters } from "types/logs";
