+++ apps/parsley/src/hooks/useQueryParam/index.ts
@@ -1,5 +1,5 @@
 import { useCallback, useMemo } from "react";
-import { conditionalToArray } from "@evg-ui/utils";
+import { conditionalToArray } from "@evg-ui/lib/utils/array";
 import { ParseOptions } from "query-string";
 import { useNavigate, useSearchParams } from "react-router-dom";
 import { QueryParams } from "constants/queryParams";
