+++ apps/spruce/lint-staged.config.js
@@ -1,4 +1,4 @@
-const baseConfig = require("../../lint-staged.base.config");
+const baseConfig = require("@evg-ui/lint-staged");
 
 module.exports = {
   ...baseConfig,
