+++ apps/spruce/src/components/TextInputWithGlyph/TextInputWithGlyph.stories.tsx
@@ -1,6 +1,6 @@
 // TODO: Relocate story to @evg-ui/components package
 import { useState } from "react";
-import { TextInputWithGlyph } from "@evg-ui/components";
+import { TextInputWithGlyph } from "@evg-ui/lib/components/TextInputWithGlyph";
 import Icon from "components/Icon";
 import { CustomStoryObj, CustomMeta } from "test_utils/types";
