--- packages/components/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@evg-ui/components",
-  "version": "0.0.0",
-  "main": "./src/index.ts",
-  "private": true,
-  "scripts": {
-    "check-types": "tsc -p tsconfig.json --noEmit",
-    "eslint:fix": "yarn eslint:strict --fix",
-    "eslint:staged": "STRICT=1 eslint",
-    "eslint:strict": "STRICT=1 eslint .",
-    "test": "vitest --typecheck=false"
-  },
-  "dependencies": {
-    "@emotion/react": "11.11.0",
-    "@emotion/styled": "11.11.0",
-    "@evg-ui/constants": "*",
-    "@leafygreen-ui/text-input": "13.0.2",
-    "@leafygreen-ui/tokens": "2.5.2"
-  },
-  "devDependencies": {
-    "@evg-ui/eslint-config": "*",
-    "eslint": "8.56.0",
-    "eslint-plugin-prettier": "5.1.3",
-    "typescript": "5.1.3",
-    "vitest": "1.6.0"
-  }
-}
