--- packages/constants/.eslintrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-  extends: ["@evg-ui"],
-  root: true,
-};
