--- packages/constants/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "baseUrl": "src"
-  }
-}
