+++ packages/lib/.eslintrc.cjs
