+++ packages/lib/lint-staged.config.js
@@ -1,0 +1,3 @@
+import baseConfig from "@evg-ui/lint-staged";
+
+export default baseConfig;
