+++ packages/lib/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@evg-ui/constants",
+  "name": "@evg-ui/lib",
   "version": "0.0.0",
-  "main": "./src/index.ts",
   "private": true,
+  "type": "module",
   "scripts": {
     "check-types": "tsc -p tsconfig.json --noEmit",
     "eslint:fix": "yarn eslint:strict --fix",
@@ -11,10 +11,14 @@
     "test": "vitest --typecheck=false"
   },
   "dependencies": {
+    "@emotion/react": "11.11.0",
+    "@emotion/styled": "11.11.0",
+    "@leafygreen-ui/text-input": "13.0.2",
     "@leafygreen-ui/tokens": "2.5.2"
   },
   "devDependencies": {
     "@evg-ui/eslint-config": "*",
+    "@evg-ui/lint-staged": "*",
     "eslint": "8.56.0",
     "eslint-plugin-prettier": "5.1.3",
     "typescript": "5.1.3",
