+++ packages/lint-staged/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@evg-ui/lint-staged",
+  "version": "0.0.0",
+  "main": "index.js",
+  "private": true
+}
