--- packages/utils/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@evg-ui/utils",
-  "version": "0.0.0",
-  "main": "./src/index.ts",
-  "private": true,
-  "scripts": {
-    "check-types": "tsc -p tsconfig.json --noEmit",
-    "eslint:fix": "yarn eslint:strict --fix",
-    "eslint:staged": "STRICT=1 eslint",
-    "eslint:strict": "STRICT=1 eslint .",
-    "test": "vitest --typecheck=false"
-  },
-  "devDependencies": {
-    "@evg-ui/eslint-config": "*",
-    "eslint": "8.56.0",
-    "eslint-plugin-prettier": "5.1.3",
-    "typescript": "5.1.3",
-    "vitest": "1.6.0"
-  }
-}
